backend updates - removed original session Structure - removed unused imports - updated weight int > float - fixed a bug where parameter name was invalid - updated frontend type names with lowercase variants.

+++ src-tauri/src/endpoints/session.rs
@@ -1,5 +1,6 @@
 use std::collections::btree_map::Range;
 use std::sync::Mutex;
+use serde::{Deserialize, Serialize};
 use tauri::webview::cookie::time::{format_description, UtcDateTime};
 use uuid::Uuid;
 use crate::{api, services, Db, SessionExercises, SessionState};
@@ -28,7 +29,7 @@ pub fn start_session(
         for i in 0..3 {
             sets.push(models::Set{
                 reps: 0,
-                weight:0,
+                weight:0.0,
                 time_completed: String::new()
             });
         }
@@ -74,12 +75,12 @@ pub fn get_session(
 }
 
 
-#[derive(Debug)]
+#[derive(Debug,Serialize,Deserialize)]
 pub struct SetUpdateDTO {
     pub exercise_id: String,
     pub set_nr: usize,
     pub reps: i32,
-    pub weight: i32,
+    pub weight: f32,
 }
 #[tauri::command]
 pub fn update_set(
